Updates to the epeebox code
Fixing a bug or 2, but there is still an issue with a double hit
where if the hits are a certain time apart, one light will go on
for the 3.5s and then the other will come on only after the first
has gone off. This needs to be sorted out.