Fixes after testing Sabre with real fencers
My understanding of Sabre was flawed, the B and C wires are actually connected
in the weapon which meant that the scoring was not working correctly as we
expected a higher ADC value. These have been updated and now Sabre should work
correctly. Now just to implement whipover and then Sabre is complete.